[Refactor] 병합 전 Commit
