+++ src/main/java/project/moseup/config/SecurityConfig.java
@@ -1,11 +1,10 @@
-package project.moseup.auth;
+package project.moseup.config;
 
 import lombok.AllArgsConstructor;
-import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -15,6 +14,8 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import project.moseup.service.MemberService;
 
+import javax.sql.DataSource;
+
 @Configuration
 @EnableWebSecurity
 @AllArgsConstructor
@@ -22,6 +23,8 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
     private MemberService memberService;
 
+    @Autowired
+    private DataSource dataSource;
     @Bean
     public PasswordEncoder passwordEncoder(){
         return new BCryptPasswordEncoder();
@@ -58,4 +61,14 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     protected void configure(AuthenticationManagerBuilder auth) throws Exception{
         auth.userDetailsService(memberService).passwordEncoder(passwordEncoder());
     }
+
+    // 유튜브 강의에서 한거
+//    @Autowired
+//    public void configureGlobal(AuthenticationManagerBuilder auth) throws  Exception{
+//        auth.jdbcAuthentication()
+//                .dataSource(dataSource)
+//                .passwordEncoder(passwordEncoder())
+//                .usersByUsernameQuery("select email, password, enabled from Member where email = ?")
+//                .authoritiesByUsernameQuery("select email, authority from authorities where email = ?");
+//    }
 }
